requirements.txt
telegram-send
RPi.GPIO